requirements.txt
pyttsx3
pyautogui
keyboard
pygame
soundfile
easyocr
tkinter
Pillow
jupyter
ipywidgets